Duplicated the Ghost AI and now they seem to have a mind of their own
Pretty creepy not gonna lie. They move when you're not looking, but then when you walk up to them, they might just stop moving, and then suddenly move again. They've also acquired the ability to pass through objects. This is nice since they are ghosts, but all of these behaviors are unintended and idk if i like them or not since idk how to control them :/